chore: move new test to folder
