+++ bouncer/commands/open_and_delete_orders.ts
@@ -8,7 +8,7 @@
 // will create 5 limit_order and then delete them all with a single extrinsic
 
 import { runWithTimeoutAndExit } from '../shared/utils';
-import { createAndDeleteMultipleOrders } from '../shared/create_and_delete_multiple_orders';
+import { createAndDeleteMultipleOrders } from '../tests/create_and_delete_multiple_orders';
 
 async function main(): Promise<void> {
   if (!process.argv[2]) {
