+++ bouncer/commands/setup_lp_account.ts
@@ -6,12 +6,15 @@
 //
 // For example: ./commands/setup_lp_account.ts //LP_3
 
-import { executeWithTimeout } from '../shared/utils';
+import { runWithTimeoutAndExit } from '../shared/utils';
 import { setupLpAccount } from '../shared/setup_lp_account';
 
 async function main() {
   const lpKey = process.argv[2];
+  if (!lpKey) {
+    throw new Error('No LP key provided');
+  }
   await setupLpAccount(lpKey);
 }
 
-await executeWithTimeout(main(), 120);
+await runWithTimeoutAndExit(main(), 120);
