+++ bouncer/test_commands/create_and_delete_multiple_orders.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env -S pnpm tsx
+import { testCancelOrdersBatch } from '../tests/create_and_delete_multiple_orders';
+
+await testCancelOrdersBatch.run();
